# arinar-v2/packages/contracts/openapi/arinar-v1.yaml
openapi: 3.1.0
info:
  title: Arinar API
  version: 1.0.0
  description: |
    Arinar V2 API - AI-native knowledge platform with multi-agent realtime discussions.
    
    **Model Provider Policy**: This system uses OpenRouter for LLM access, not direct OpenAI/Anthropic integrations.
  contact:
    name: Arinar Platform Team
servers:
  - url: http://localhost:8000
    description: Local development
  - url: https://api.arinar.dev
    description: Development environment
  - url: https://api.arinar.com
    description: Production
...
paths:
...
  /debates/{debate_id}/events/stream:
    get:
      operationId: streamDebateEvents
      summary: Stream debate events via SSE (M2)
      description: |
        Real-time Server-Sent Events stream of debate events.
        Stream automatically closes when debate ends.
      tags:
        - events
      parameters:
        - $ref: '#/components/parameters/DebateIdParam'
        - name: since
          in: query
          description: Start streaming from this sequence number
          required: false
          schema:
            type: integer
            minimum: 0
      responses:
        '200':
          description: SSE stream established
          content:
            text/event-stream:
              schema:
                type: string
                description: Server-Sent Events stream with debate_event, state_update, and stream_end messages
        '404':
          description: Debate not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /openrouter/models:
    get:
      operationId: listOpenRouterModels
      summary: Fetch OpenRouter model catalog
      description: |
        Fetches available models from OpenRouter API using user's BYOK key.
        Key is never stored. Results cached 60s.
      tags:
        - openrouter
      parameters:
        - name: X-OpenRouter-Key
          in: header
          required: true
          schema:
            type: string
          description: "OpenRouter API key (BYOK)"
      responses:
        '200':
          description: Model list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ModelListResponse'
        '400':
          description: Missing API key
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Invalid API key
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /openrouter/account:
    get:
      operationId: getOpenRouterAccount
      summary: Get OpenRouter account info
      description: |
        Fetches account details from OpenRouter: usage, limits, credits.
        Key is never stored.
      tags:
        - openrouter
      parameters:
        - name: X-OpenRouter-Key
          in: header
          required: true
          schema:
            type: string
          description: "OpenRouter API key (BYOK)"
      responses:
        '200':
          description: Account information
          content:
            application/json:
              schema:
                type: object
                properties:
                  key:
                    type: object
                    description: Key usage and limits
                  credits:
                    type: object
                    nullable: true
                    description: Credits info (null if not available)
                  note:
                    type: string
                    nullable: true
                    description: Additional info or warnings
        '400':
          description: Missing API key
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Invalid API key
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  parameters:
    DebateIdParam:
      name: debate_id
      in: path
      description: Debate session UUID
      required: true
      schema:
        type: string
        format: uuid
  schemas:
...
    ErrorResponse:
      type: object
      required:
        - error
        - message
      properties:
        error:
          type: string
        message:
          type: string
        details:
          type: object
        timestamp:
          type: string
          format: date-time
    ModelListResponse:
      type: object
      required:
        - models
      properties:
        models:
          type: array
          items:
            $ref: '#/components/schemas/OpenRouterModel'
        cached:
          type: boolean
          default: false
    OpenRouterModel:
      type: object
      required:
        - id
        - name
      properties:
        id:
          type: string
        name:
          type: string
        context_length:
          type: integer
        pricing:
          type: object
          additionalProperties: true
        description:
          type: string
...